refactor: move regex to separated file, remove uuid regex
