+++ src/plugins/typebox.ts
@@ -21,7 +21,7 @@ import {
 
 import { MAX_ITEM_NAME_LENGTH, MAX_USERNAME_LENGTH, MIN_USERNAME_LENGTH } from '@graasp/sdk';
 
-import { ITEM_NAME_REGEX } from '../schemas/global';
+import { ITEM_NAME_REGEX } from '../schemas/regex';
 import { discriminable } from './typebox/discriminable';
 
 /**
