+++ src/schemas/regex.ts
@@ -1,0 +1,3 @@
+export const ITEM_NAME_REGEX = '^\\S+( \\S+)*$';
+// allow empty strings or words separated by one space
+export const EMPTY_OR_SPACED_WORDS_REGEX = /^(\S+( \S+)*)?$/;
