+++ src/services/itemMembership/schemas.ts
@@ -6,7 +6,7 @@ import { FastifySchema } from 'fastify';
 import { PermissionLevel } from '@graasp/sdk';
 
 import { customType, registerSchemaAsRef } from '../../plugins/typebox';
-import { UUID_REGEX, entityIdSchemaRef, errorSchemaRef } from '../../schemas/global';
+import { entityIdSchemaRef, errorSchemaRef } from '../../schemas/global';
 import { augmentedAccountSchemaRef, nullableAugmentedAccountSchemaRef } from '../account/schemas';
 import { itemIdSchemaRef, itemSchemaRef } from '../item/schemas';
 
@@ -87,7 +87,7 @@ export const getItems = {
   ),
   response: {
     [StatusCodes.OK]: Type.Object({
-      data: Type.Record(Type.String({ pattern: UUID_REGEX }), Type.Array(itemMembershipSchemaRef)),
+      data: Type.Record(Type.String({ format: 'uuid' }), Type.Array(itemMembershipSchemaRef)),
       errors: Type.Array(errorSchemaRef),
     }),
   },
